pyproject.toml
[tool.poetry]
name = "edfi_amt_data_lake"
version = "0.1.0"
description = "API Data converted into a Data Lake"
authors = ["EdFi"]
license = "Apache License 2.0"

[tool.poetry.dependencies]
python = "^3.8"
dagit = "0.15.0"
dagster = "0.15.0"
dagster-graphql = "0.15.0"
python-dotenv = "0.20.0"
requests = "2.28.0"
tqdm = "4.64.0"
uvicorn = "0.17.6"
python-decouple = "^3.6"
pandas = "^1.4.3"
fastparquet = "^0.8.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pytest-cov = "^2.10.1"
autoflake = "^1.4"
flake8 = "^3.8.4"
mypy = "^0.790"
isort = "^5.0"
pre-commit = "^2.8.2"
black = "^20.8b1"
pydantic = "^1.10.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
